# Typecheck rewrite