--- a/lib/graphql.rb
+++ b/lib/graphql.rb
@@ -69,6 +69,8 @@ require "graphql/invalid_null_error"
 require "graphql/unresolved_type_error"
 require "graphql/query"
 require "graphql/internal_representation"
+# TODO: rename
+require "graphql/static_analysis"
 require "graphql/static_validation"
 require "graphql/version"
 require "graphql/relay"