--- a/lib/graphql/input_object_type.rb
+++ b/lib/graphql/input_object_type.rb
@@ -101,5 +101,11 @@ module GraphQL
 
       result
     end
+
+    # @param [String] Name for an argument on this input
+    # @return [GraphQL::Argument, nil] The argument for that name, if it's defined
+    def get_argument(name)
+      arguments[name]
+    end
   end
 end