--- a/lib/graphql/static_analysis/type_check/any_field.rb
+++ b/lib/graphql/static_analysis/type_check/any_field.rb
@@ -15,6 +15,10 @@ module GraphQL
         def arguments
           ARGUMENTS
         end
+
+        def get_argument(name)
+          AnyArgument
+        end
       end
     end
   end