@rmosolgo 2016-10-05
Just finished reimplementing the last rule, naturally, I couldn't resist a preliminary benchmark. TL;DR: 2x faster :tada: !

```ruby 
# validation_benchmark.rb
require "graphql"
require "benchmark/ips"
require "support/dairy_data"
require "support/dairy_app"
document = GraphQL.parse(GraphQL::Introspection::INTROSPECTION_QUERY)
schema = DummySchema
query = GraphQL::Query.new(schema, document: document)

Benchmark.ips do |x|
  x.report("old validation") {
    validator = GraphQL::StaticValidation::Validator.new(schema: schema)
    validator.validate(query)
  }

  x.report("new validation") {
    visitor = GraphQL::Language::Visitor.new(document)
    analysis = GraphQL::StaticAnalysis.prepare(visitor, schema: schema)
    visitor.visit
    analysis.errors
  }
  x.compare!
end
```

```sh
~/code/graphql $ ruby -Ilib -Ispec validation_benchmark.rb
Warming up --------------------------------------
      old validation    43.000  i/100ms
      new validation   104.000  i/100ms
Calculating -------------------------------------
      old validation    430.374  (± 4.0%) i/s -      2.193k in   5.104203s
      new validation      1.059k (± 3.5%) i/s -      5.304k in   5.012636s

Comparison:
      new validation:     1059.5 i/s
      old validation:      430.4 i/s - 2.46x  slower
```