@rmosolgo 2016-10-05
I guess it's not quite fair: I haven't re-implemented `queryPath` yet!