@rmosolgo 2017-02-04
Some of the lessons-learned from this branch have been merged into master. The performance better and I've removed the ad-hoc recursion from the rules. There are still some goals unaccomplished here (eg, improved handling of undefined types and fields) but the main issues are resolved!